Refactored existing tests to integration project #840
