+++ tests/Test.ADAL.NET.Integration/Test.ADAL.NET.Integration.csproj
@@ -81,6 +81,7 @@
   <ItemGroup>
     <Compile Include="AcquireTokenSilentTests.cs" />
     <Compile Include="DeviceProfileTests.cs" />
+    <Compile Include="ManagedUsercredentialTests.cs" />
     <Compile Include="PromptBehaviorTests.cs" />
     <Compile Include="SovereignCloudsTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
